requirements.txt
black==24.4.2
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
filelock==3.14.0
flake8==7.0.0
identify==2.5.36
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pre-commit==3.7.1
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyYAML==6.0.1
virtualenv==20.26.2
